+++ patzilla-ui/navigator/app/document/index.js
@@ -1,3 +1,3 @@
 require('./document-controller.js');
 require('./document-view.js');
-require('./family.js');
+require('./biblio-family.js');